Именил CMakeLists.txt

+++ core.cpp
@@ -8,7 +8,7 @@ int main()
 {
 	const size_t ARR_SIZE = 10;
 	String * arr = new String[ARR_SIZE];
-	for(int i = 0; i < ARR_SIZE; i++)
+	for(size_t i = 0; i < ARR_SIZE; i++)
 		arr[i] = "Hello World \n";
 	for(size_t i = 0; i < ARR_SIZE; i++)
 		arr[i][11] = i + '0';
